requirements.txt
# Python 3.10+ Required (3.11 recommended for performance)
# Tested with Python 3.10, 3.11, and 3.12
# Note: Python 3.13 may have compatibility issues with some packages

# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.9.0
click==8.1.7
aiohttp>=3.9.3
aiofiles>=23.2.1

# Salad Cloud Transcription SDK (replacing Whisper)
salad-cloud-transcription-sdk==1.0.0a1

# OpenAI for LLM enrichment
openai>=1.12.0

# Anthropic (optional for enrichment)
anthropic>=0.18.0

# RingCentral SDK
ringcentral>=0.7.9

# Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Google Drive
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Google Vertex AI RAG
google-cloud-aiplatform>=1.71.0
google-cloud-storage>=2.0.0

# Database and state management
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Scheduling and monitoring
schedule==1.2.0
APScheduler==3.10.4
prometheus-client==0.19.0
psutil==5.9.6

# Logging and error tracking
python-json-logger==2.0.7
# sentry-sdk==1.39.0  # Optional - uncomment if using Sentry

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.3

# Additional utilities
tenacity==8.2.3  # For retry logic
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1  # For configuration files

# Development tools (optional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1
types-requests==2.31.0